+++ lib/index.ts
@@ -1,0 +1,9 @@
+export { AutoSizer } from "./components/AutoSizer";
+
+export type {
+  AutoSizerProps,
+  AutoSizerBox,
+  AutoSizerChildProps
+} from "./components/types";
+
+export type { Size } from "./types";