+++ scripts/compile-docs.ts
@@ -1,0 +1,10 @@
+import { compileComponents } from "./utils/docs/compileComponents.ts";
+
+async function run() {
+  await compileComponents({
+    componentNames: ["AutoSizer.tsx"],
+    outputDirName: "js-docs"
+  });
+}
+
+run();