+++ src/index.ts
@@ -1,5 +1,0 @@
-import { AutoSizer } from "./AutoSizer";
-
-export default AutoSizer;
-
-export * from "./types";